+++ python/evalica/naive.py
@@ -2,7 +2,7 @@ import numpy as np
 import numpy.typing as npt
 
 
-def bradley_terry(M: npt.NDArray[np.int64], tolerance: float = 1e-8) -> tuple[  # noqa: N803
+def bradley_terry(M: npt.NDArray[np.float64], tolerance: float = 1e-8) -> tuple[  # noqa: N803
     npt.NDArray[np.float64], int]:
     T = M.T + M  # noqa: N806
     active = T > 0
